fix wait to use the correct wait index
Waiting on a key will fail if the wait index is too old. This happens
when waiting on a key that does not change a lot but other keys are
changing. Previous behavior would fail with an error which the caller
would have to handle. This commit internalizes the logic to this module.

Alternative approach is to document this error and make etcdjs.wait
callers handle it.